+++ rust/main/chains/hyperlane-cosmos/src/error.rs
@@ -29,8 +29,8 @@ pub enum HyperlaneCosmosError {
     /// Cosmos error report
     #[error("{0}")]
     CosmosErrorReport(#[from] Box<cosmrs::ErrorReport>),
-    #[error("{0}")]
     /// Cosmrs Tendermint Error
+    #[error("{0}")]
     CosmrsTendermintError(#[from] Box<cosmrs::tendermint::Error>),
     #[error("{0}")]
     /// CosmWasm Error
@@ -79,6 +79,30 @@ pub enum HyperlaneCosmosError {
     ParsingAttemptsFailed(Vec<HyperlaneCosmosError>),
 }
 
+impl From<cosmrs::ErrorReport> for HyperlaneCosmosError {
+    fn from(value: cosmrs::ErrorReport) -> Self {
+        HyperlaneCosmosError::CosmosErrorReport(Box::new(value))
+    }
+}
+
+impl From<tonic::Status> for HyperlaneCosmosError {
+    fn from(value: tonic::Status) -> Self {
+        HyperlaneCosmosError::GrpcError(Box::new(value))
+    }
+}
+
+impl From<cosmrs::Error> for HyperlaneCosmosError {
+    fn from(value: cosmrs::Error) -> Self {
+        HyperlaneCosmosError::CosmosError(Box::new(value))
+    }
+}
+
+impl From<cosmwasm_std::StdError> for HyperlaneCosmosError {
+    fn from(value: cosmwasm_std::StdError) -> Self {
+        HyperlaneCosmosError::CosmWasmError(Box::new(value))
+    }
+}
+
 impl From<HyperlaneCosmosError> for ChainCommunicationError {
     fn from(value: HyperlaneCosmosError) -> Self {
         ChainCommunicationError::from_other(value)
@@ -94,5 +118,17 @@ impl From<PublicKeyError> for HyperlaneCosmosError {
 impl From<cometbft_rpc::Error> for HyperlaneCosmosError {
     fn from(value: cometbft_rpc::Error) -> Self {
         HyperlaneCosmosError::CometbftRpcError(Box::new(value))
+    }
+}
+
+impl From<cometbft::Error> for HyperlaneCosmosError {
+    fn from(value: cometbft::Error) -> Self {
+        HyperlaneCosmosError::CometbftError(Box::new(value))
+    }
+}
+
+impl From<cosmrs::tendermint::Error> for HyperlaneCosmosError {
+    fn from(value: cosmrs::tendermint::Error) -> Self {
+        HyperlaneCosmosError::CosmrsTendermintError(Box::new(value))
     }
 }