+++ typescript/widgets/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @hyperlane-xyz/widgets
 
+## 18.0.0
+
+### Patch Changes
+
+- Updated dependencies [552b253b9]
+- Updated dependencies [ba832828f]
+- Updated dependencies [cfc0eb2a7]
+  - @hyperlane-xyz/sdk@18.0.0
+  - @hyperlane-xyz/utils@18.0.0
+  - @hyperlane-xyz/cosmos-sdk@18.0.0
+
 ## 17.0.0
 
 ### Major Changes